+++ cpp/arcae/isolated_table_proxy.cc
@@ -5,11 +5,10 @@
 #include <limits>
 #include <memory>
 
+#include <arrow/status.h>
+#include <arrow/util/future.h>
 #include <arrow/util/logging.h>
-#include "arrow/status.h"
-#include "arrow/util/functional.h"
-#include "arrow/util/future.h"
-#include "arrow/util/thread_pool.h"
+#include <arrow/util/thread_pool.h>
 
 #include <casacore/tables/Tables/TableProxy.h>